fix: missed some int->float for ray_angle

+++ maze/src/check_horizontal_intersections.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include <math.h>
 
-int check_horizontal_intersections(int *horizontal_coords, int ray_angle,
+int check_horizontal_intersections(int *horizontal_coords, float ray_angle,
     Player_POV *player, char (*map)[MAP_WIDTH])
 {
     int start_x;
@@ -21,13 +21,13 @@ int check_horizontal_intersections(int *horizontal_coords, int ray_angle,
         printf("Player x: %i\n", player->x_coord);
         printf("Player y: %i\n", player->y_coord);
         printf("Player angle: %i\n", player->angle);
-        printf("ray angle: %i\n", ray_angle);
+        printf("ray angle: %f\n", ray_angle);
         printf("horizontal coords x: %i\n", horizontal_coords[0]);
         printf("horizontal coords y: %i\n", horizontal_coords[1]);
         printf("map[0][0] %c\n", map[0][0]);
     } else if (DEBUG) {
         printf("%s\n", "====== START CHECK_HORIZONTAL_INTERSECTIONS =====");
-        printf("ray angle: %i\n", ray_angle);
+        printf("ray angle: %f\n", ray_angle);
     }
 
     if (ray_horizontal)
@@ -44,7 +44,7 @@ int check_horizontal_intersections(int *horizontal_coords, int ray_angle,
             * CUBE_LENGTH + CUBE_LENGTH;
         start_x = player->x_coord + (start_y - player->y_coord)/tan(ray_angle);
     } else {
-        printf("warning! ray_angle %i is over 360 degrees, did not handle the case\n", ray_angle);
+        printf("warning! ray_angle %f is over 360 degrees, did not handle the case\n", ray_angle);
         return 0;
     }
